+++ org.knime.base/src/org/knime/base/node/mine/regression/linear2/learner/Learner.java
@@ -154,8 +154,8 @@ final class Learner extends RegressionStatisticsLearner {
     @Override
     protected void processTable(final ExecutionMonitor exec, final RegressionTrainingData trainingData,
         final SummaryStatistics[] stats, final UpdatingMultipleLinearRegression regr) throws CanceledExecutionException {
-        int r = 1;
-        final int rowCount = trainingData.getRowCount();
+        long r = 1;
+        final long rowCount = trainingData.getRowCount();
         for (RegressionTrainingRow row : trainingData) {
             exec.checkCanceled();
             if (!row.hasMissingCells()) {
